moved grid_pixels_2d_slim_from
