+++ test_autoarray/geometry/test_geometry_util.py
@@ -1079,3 +1079,56 @@ def test__transform_2d_grid_from_reference_frame():
     )
 
     assert grid_2d == pytest.approx(original_grid_2d, 1.0e-4)
+
+
+def test__grid_pixels_2d_slim_from():
+
+    # coordinates in centres_of_pixels
+
+    grid_scaled = np.array([[1.0, -2.0], [1.0, 2.0], [-1.0, -2.0], [-1.0, 2.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixels_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (
+        grid_pixels == np.array([[0.5, 0.5], [0.5, 1.5], [1.5, 0.5], [1.5, 1.5]])
+    ).all()
+
+    # coordinates top-left of pixels
+
+    grid_scaled = np.array([[2.0, -4], [2.0, 0.0], [0.0, -4], [0.0, 0.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixels_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
+
+    # coordinates bottom-right of pixels
+
+    grid_scaled = np.array([[0.0, 0.0], [0.0, 4.0], [-2.0, 0.0], [-2.0, 4.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixels_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([[1, 1], [1, 2], [2, 1], [2, 2]])).all()
+
+    # -1.0 from all entries for a origin of (-1.0, -1.0)
+    grid_scaled = np.array([[-1.0, -1.0], [-1.0, 3.0], [-3.0, -1.0], [-3.0, 3.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixels_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+        origin=(-1.0, -1.0),
+    )
+
+    assert (grid_pixels == np.array([[1, 1], [1, 2], [2, 1], [2, 2]])).all()
